+++ o365spray/core/handlers/sprayer/modules/base.py
@@ -21,6 +21,7 @@ from o365spray.core.utils import (
     text_colors,
 )
 
+
 urllib3.disable_warnings(urllib3.exceptions.InsecureRequestWarning)
 
 
@@ -76,6 +77,8 @@ class SprayerBase(BaseHandler):
             ValueError: if no output directory provided when output writing
               is enabled
         """
+        super().__init__(*args, **kwargs)
+
         if writer and not output_dir:
             raise ValueError("Missing 1 required argument: 'output_dir'")